liveness: Move TestNodeLivenessLivenessStatus to liveness
Previously this method was in the storepool package, however it makes
more sense to have this in the liveness package as it is testing the
functionality there.

Epic: none

Release note: None
