+++ pkg/kv/kvserver/allocator/storepool/store_pool_test.go
@@ -862,154 +862,3 @@ func TestStorePoolDecommissioningReplicas(t *testing.T) {
 		t.Fatalf("expected decommissioning replicas %+v; got %+v", e, a)
 	}
 }
-
-func TestNodeLivenessLivenessStatus(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-	now := hlc.Timestamp{
-		WallTime: time.Date(2020, 0, 0, 0, 0, 0, 0, time.UTC).UnixNano(),
-	}
-	threshold := 5 * time.Minute
-
-	for _, tc := range []struct {
-		name     string
-		liveness livenesspb.Liveness
-		expected livenesspb.NodeLivenessStatus
-	}{
-		// Valid status.
-		{
-			name: "Valid 5 min",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(5 * time.Minute).ToLegacyTimestamp(),
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_LIVE,
-		},
-		{
-			name: "Expires just slightly in the future",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(time.Nanosecond).ToLegacyTimestamp(),
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_LIVE,
-		},
-		{
-			name: "Expired status",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.ToLegacyTimestamp(),
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
-		},
-		{
-			name: "Max bound of expired",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(-threshold).AddDuration(time.Nanosecond).ToLegacyTimestamp(),
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
-		},
-		{
-			name: "Dead",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_DEAD,
-		},
-		{
-			name: "Decommissioning",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(time.Second).ToLegacyTimestamp(),
-				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONING,
-		},
-		{
-			name: "Decommissioning + expired",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
-				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONED,
-		},
-		{
-			name: "Decommissioning + live",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(time.Second).ToLegacyTimestamp(),
-				Membership: livenesspb.MembershipStatus_DECOMMISSIONED,
-				Draining:   false,
-			},
-			// Despite having marked the node as fully decommissioned, through
-			// this NodeLivenessStatus API we still surface the node as
-			// "Decommissioning". See #50707 for more details.
-			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONING,
-		},
-		{
-			name: "Decommissioning + expired",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
-				Membership: livenesspb.MembershipStatus_DECOMMISSIONED,
-				Draining:   false,
-			},
-			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONED,
-		},
-		{
-			name: "Draining",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.AddDuration(5 * time.Minute).ToLegacyTimestamp(),
-				Draining:   true,
-			},
-			expected: livenesspb.NodeLivenessStatus_DRAINING,
-		},
-		{
-			name: "Decommissioning that is unavailable",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.ToLegacyTimestamp(),
-				Draining:   false,
-				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
-			},
-			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
-		},
-		{
-			name: "Draining that is unavailable",
-			liveness: livenesspb.Liveness{
-				NodeID:     1,
-				Epoch:      1,
-				Expiration: now.ToLegacyTimestamp(),
-				Draining:   true,
-			},
-			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			if a, e := LivenessStatus(tc.liveness, now, threshold), tc.expected; a != e {
-				t.Errorf("liveness status was %s, wanted %s for %+v, %v", a.String(), e.String(), tc.liveness, now)
-			}
-		})
-	}
-}
