+++ pkg/kv/kvserver/liveness/cache.go
@@ -64,19 +64,22 @@ func newCache(
 
 	c.notifyLivenessChanged = cbFn
 
-	// NB: we should consider moving this registration to .Start() once we
-	// have ensured that nobody uses the server's KV client (kv.DB) before
-	// nl.Start() is invoked. At the time of writing this invariant does
-	// not hold (which is a problem, since the node itself won't be live
-	// at this point, and requests routed to it will hang).
-	livenessRegex := gossip.MakePrefixPattern(gossip.KeyNodeLivenessPrefix)
-	c.gossip.RegisterCallback(livenessRegex, c.livenessGossipUpdate)
+	// Gossip is nil in some tests.
+	if c.gossip != nil {
+		// NB: we should consider moving this registration to .Start() once we
+		// have ensured that nobody uses the server's KV client (kv.DB) before
+		// nl.Start() is invoked. At the time of writing this invariant does
+		// not hold (which is a problem, since the node itself won't be live
+		// at this point, and requests routed to it will hang).
+		livenessRegex := gossip.MakePrefixPattern(gossip.KeyNodeLivenessPrefix)
+		c.gossip.RegisterCallback(livenessRegex, c.livenessGossipUpdate)
 
-	// Enable redundant callbacks for the store keys because we use these
-	// callbacks as a clock to determine when a store was last updated even if it
-	// hasn't otherwise changed.
-	storeRegex := gossip.MakePrefixPattern(gossip.KeyStoreDescPrefix)
-	c.gossip.RegisterCallback(storeRegex, c.storeGossipUpdate, gossip.Redundant)
+		// Enable redundant callbacks for the store keys because we use these
+		// callbacks as a clock to determine when a store was last updated even if it
+		// hasn't otherwise changed.
+		storeRegex := gossip.MakePrefixPattern(gossip.KeyStoreDescPrefix)
+		c.gossip.RegisterCallback(storeRegex, c.storeGossipUpdate, gossip.Redundant)
+	}
 	return &c
 }
