+++ pkg/kv/kvserver/liveness/liveness_test.go
@@ -13,12 +13,16 @@ package liveness
 import (
 	"fmt"
 	"testing"
+	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness/livenesspb"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
+	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 )
 
 func TestShouldReplaceLiveness(t *testing.T) {
@@ -127,6 +131,165 @@ func TestShouldReplaceLiveness(t *testing.T) {
 		t.Run("", func(t *testing.T) {
 			if act := livenessChanged(test.old, test.new); act != test.exp {
 				t.Errorf("unexpected update: %+v", test)
+			}
+		})
+	}
+}
+
+func TestNodeLivenessLivenessStatus(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	defer log.Scope(t).Close(t)
+	clock := hlc.NewClockForTesting(timeutil.NewManualTime(timeutil.Unix(100000, 100000)))
+	nl := NewNodeLiveness(NodeLivenessOptions{Clock: clock, Settings: cluster.MakeTestingClusterSettings()})
+	now := clock.Now()
+	threshold := TimeUntilNodeDead.Default()
+
+	for _, tc := range []struct {
+		name     string
+		liveness livenesspb.Liveness
+		expected livenesspb.NodeLivenessStatus
+	}{
+		// Valid status.
+		{
+			name: "Valid 5 min",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(5 * time.Minute).ToLegacyTimestamp(),
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_LIVE,
+		},
+		{
+			name: "Expires just slightly in the future",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(time.Nanosecond).ToLegacyTimestamp(),
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_LIVE,
+		},
+		{
+			name: "Expired status",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.ToLegacyTimestamp(),
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
+		},
+		{
+			name: "Max bound of expired",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(-threshold).AddDuration(time.Nanosecond).ToLegacyTimestamp(),
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
+		},
+		{
+			name: "Dead",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_DEAD,
+		},
+		{
+			name: "Decommissioning",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(time.Second).ToLegacyTimestamp(),
+				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONING,
+		},
+		{
+			name: "Decommissioning + expired",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
+				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONED,
+		},
+		{
+			name: "Decommissioning + live",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(time.Second).ToLegacyTimestamp(),
+				Membership: livenesspb.MembershipStatus_DECOMMISSIONED,
+				Draining:   false,
+			},
+			// Despite having marked the node as fully decommissioned, through
+			// this NodeLivenessStatus API we still surface the node as
+			// "Decommissioning". See #50707 for more details.
+			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONING,
+		},
+		{
+			name: "Decommissioning + expired",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(-threshold).ToLegacyTimestamp(),
+				Membership: livenesspb.MembershipStatus_DECOMMISSIONED,
+				Draining:   false,
+			},
+			expected: livenesspb.NodeLivenessStatus_DECOMMISSIONED,
+		},
+		{
+			name: "Draining",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.AddDuration(5 * time.Minute).ToLegacyTimestamp(),
+				Draining:   true,
+			},
+			expected: livenesspb.NodeLivenessStatus_DRAINING,
+		},
+		{
+			name: "Decommissioning that is unavailable",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.ToLegacyTimestamp(),
+				Draining:   false,
+				Membership: livenesspb.MembershipStatus_DECOMMISSIONING,
+			},
+			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
+		},
+		{
+			name: "Draining that is unavailable",
+			liveness: livenesspb.Liveness{
+				NodeID:     1,
+				Epoch:      1,
+				Expiration: now.ToLegacyTimestamp(),
+				Draining:   true,
+			},
+			expected: livenesspb.NodeLivenessStatus_UNAVAILABLE,
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			// TODO: Add additional testing around the lastNodeUpdate.
+			// Pretend that the node descriptor has been recently updated so we don't
+			// fail due to that.
+			nl.cache.mu.lastNodeUpdate = make(map[roachpb.NodeID]UpdateInfo)
+			nl.cache.mu.lastNodeUpdate[1] = UpdateInfo{lastUpdateTime: now}
+
+			nv := nl.convertToNodeVitality(tc.liveness)
+			ls := nv.LivenessStatus()
+			if ls != tc.expected {
+				t.Errorf("liveness status was %s, expected %s for %+v at %v", ls, tc.expected, tc.liveness, now)
 			}
 		})
 	}
